Reorganize tests + test commons
